DOS-CRLF fixes: decode \r\n within quotes, add fix output \r\n
Split the \r\n output into a separate functor that can be plugged
into the output chain at compile time.

This will add \r to any \n (even those inside quotes.

Note that it will detect \r\n in the datastream, and will not
output a newline if there was already a \r there.