Added option to allow NULL chars (default: YES)
We used to disallow null characters, but technically there is no reason
to allow them, and sometimes we need to be able to load weird CSV files
with nulls.

I've set the default to TRUE (allow nulls), because the only time we've
run into issues is either when the file has been corrupted (eg truncated),
or when we are importing a file with a null in some CSV fields (bad
writer?), and we would want to import and handle downstream anyway.